chore: rename eleventy config folder
